+++ libs/deck-loaders/src/vessels/lib/parse-tracks.ts
@@ -1,6 +1,6 @@
 import { TrackField } from '@globalfishingwatch/api-types'
 import { VesselTrackData } from './types'
-import { deckTrackDecoder } from './vessel-track-proto'
+import { DeckTrack } from './vessel-track-proto'
 
 export const DEFAULT_NULL_VALUE = -Math.pow(2, 31)
 
@@ -98,5 +98,18 @@ export const trackValueArrayToSegments = (valueArray: number[], fields_: TrackFi
 }
 
 export const parseTrack = (arrayBuffer: ArrayBuffer): VesselTrackData => {
-  return deckTrackDecoder(arrayBuffer)
+  const track = DeckTrack.decode(new Uint8Array(arrayBuffer)) as any
+  return {
+    ...track,
+    attributes: {
+      getPath: {
+        value: new Float32Array(track.attributes.getPath.value),
+        size: track.attributes.getPath.size,
+      },
+      getTimestamp: {
+        value: new Float32Array(track.attributes.getTimestamp.value),
+        size: track.attributes.getTimestamp.size,
+      },
+    },
+  } as VesselTrackData
 }
